Addressed all the comments made by Russ in his review of 2790.  The point about PipeQA being vulnerable to SQL injection will be addressed in a new ticket, 2897. Basic changes:
DbQaData: Rewritten code that assembled the parts of the SQL query that go in the WHERE clause.
The usePoly section was a bit of hardwired code that Steve Bickerton inserted in order to provide an example of a polygon-based search when we migrate to qserv.
Documented usePoly code.
Rewrote how pieces of SQL query were assembled for style.

QaDataUtils: Removed mutable objects from default parameters here and in other places.

PhotCompareQaTask and other files:
Style change made to statements that compare objects to None, to improve readability.